feat: Created SupabaseStreamBuilder class and changed the API
