+++ lib/src/supabase.dart
@@ -4,6 +4,7 @@ import 'package:gotrue/gotrue.dart';
 import 'package:postgrest/postgrest.dart';
 import 'package:realtime_client/realtime_client.dart';
 import 'package:storage_client/storage_client.dart';
+import 'package:supabase/src/supabase_stream_builder.dart';
 
 import 'supabase_query_builder.dart';
